Переделал отрисовку main страницы без карточек по человечески
