--- project/src/components/cities-empty/cities-empty.tsx
@@ -1,29 +1,0 @@
-import { useEffect } from 'react';
-
-function CitiesEmpty(): JSX.Element {
-  useEffect(() => {
-    const page: HTMLElement | null = document.querySelector('.page__main');
-    (page as HTMLElement).classList.add('page__main--index-empty');
-    return () => {
-      (page as HTMLElement).classList.remove('page__main--index-empty');
-    };
-  }, []);
-
-  return (
-    <div className="cities">
-      <div className="cities__places-container cities__places-container--empty container">
-        <section className="cities__no-places">
-          <div className="cities__status-wrapper tabs__content">
-            <b className="cities__status">No places to stay available</b>
-            <p className="cities__status-description">
-              We could not find any property available at the moment in
-              Dusseldorf
-            </p>
-          </div>
-        </section>
-        <div className="cities__right-section"></div>
-      </div>
-    </div>
-  );
-}
-export default CitiesEmpty;
