+++ project/src/pages/main.tsx
@@ -1,5 +1,6 @@
 import { useEffect } from 'react';
 import Header from '../components/header/header';
+import MainScreenEmpty from '../components/main-screen-empty/main-screen-empty';
 import MainScreen from '../components/main-screen/main-screen';
 import { useAppDispatch, useAppSelector } from '../hooks';
 import { fetchOffersAction } from '../store/api-actions';
@@ -10,11 +11,12 @@ function Main(): JSX.Element {
   useEffect(() => {
     dispatch(fetchOffersAction());
   }, [isOffersLoaded]);
+  const offers = useAppSelector(({ MAIN }) => MAIN.offers);
 
   return (
     <div className="page page--gray page--main">
       <Header />
-      <MainScreen />
+      {isOffersLoaded && offers.length === 0 ? <MainScreenEmpty /> : <MainScreen/>}
     </div>
   );
 }
